Refactor code for main and books creation and listing
