+++ app.rb
@@ -3,60 +3,15 @@ require_relative './rental'
 require_relative './student'
 require_relative './teacher'
 require_relative './book'
+require_relative './book_logic'
 
 class App
-  attr_reader :people, :books, :rentals
+  attr_accessor :people, :books_class, :rentals
 
   def initialize
     @people = []
-    @books = []
+    @books_class = BookLogic.new
     @rentals = []
-  end
-
-  def options
-    puts 'Please choose an option by enterin a number :'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person id'
-    puts "7 - Exit\n\n"
-    gets.chomp
-  end
-
-  def run
-    choose = options
-    case choose
-    when '1'
-      list_all_books
-    when '2'
-      list_all_people
-    when '3'
-      create_person
-    when '4'
-      create_book
-    when '5'
-      create_rental
-    when '6'
-      list_all_rentals
-    else
-      puts 'Thank You for using the app.'
-      exit
-    end
-  end
-
-  def list_all_books
-    if @books.size.positive?
-      puts 'Here are the books registered :'
-      @books.each_with_index do |book, index|
-        puts "#{index}- Title: \"#{book.title}\" , Author: #{book.author} "
-      end
-    else
-      puts 'Sorry No book registered. Try after adding a book'
-    end
-    gets.chomp
-    run
   end
 
   def list_all_people
@@ -69,18 +24,6 @@ class App
       puts 'There are no people registered'
     end
     gets.chomp
-    run
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp.capitalize
-    print 'Author: '
-    author = gets.chomp.capitalize
-    book = Book.new(title, author)
-    @books.push(book)
-    puts 'Book created successfully'
-    run
   end
 
   def create_person
@@ -94,7 +37,6 @@ class App
     end
     puts 'Person created successfully'
     gets
-    run
   end
 
   def yes_no(input)
