+++ data_values.rb
@@ -1,0 +1,9 @@
+module Data
+    
+    @books = []
+
+    def self.books
+        @books
+    end
+
+end
